Added Makefile -- win64

+++ Example_Programs/calculator_complex/src/complex_div.c
@@ -8,12 +8,12 @@ error_t complex_div(complex_t* s1, complex_t* s2, complex_t* res){
         return  ERROR_DIV_BY_ZERO;
     }
 
-    int a=(((s1->real)*(s2->real))+((s1->imaginary)*(s2->imaginary)))/(pow(s2->real,2)+pow(s2->imaginary,2));
-    int b=(((s2->real)*(s1->imaginary))-((s1->real)*(s2->imaginary)))/(pow(s2->real,2)+pow(s2->imaginary,2));
+    float a=(((s1->real)*(s2->real))+((s1->imaginary)*(s2->imaginary)))/(pow(s2->real,2)+pow(s2->imaginary,2));
+    float b=(((s2->real)*(s1->imaginary))-((s1->real)*(s2->imaginary)))/(pow(s2->real,2)+pow(s2->imaginary,2));
 
     res->real = a;
     res->imaginary = b;
-    printf("The division is %d + i%d\n", a,b);
+    printf("The division is %f + i%f\n", a,b);
     return SUCCESS;
 
 }
