+++ Example_Programs/calculator_complex/src/complex_mul.c
@@ -2,11 +2,11 @@
 #include "complex_calculator.h"
 
 error_t complex_mul(complex_t* s1, complex_t* s2, complex_t* res){
-    int a=((s1->real)*(s2->real))-((s1->imaginary)*(s2->imaginary));
-    int b=((s1->real)*(s2->imaginary))+((s2->real)*(s1->imaginary));
+    float a=((s1->real)*(s2->real))-((s1->imaginary)*(s2->imaginary));
+    float b=((s1->real)*(s2->imaginary))+((s2->real)*(s1->imaginary));
     res->real = a;
     res->imaginary = b;
-    printf("The mutiplication is %d + i%d\n", a,b);
+    printf("The mutiplication is %f + i%f\n", a,b);
     return SUCCESS;
 
 }
